+++ CHANGELOG.md
@@ -4,7 +4,18 @@ This changelog follows the specifications detailed in: [Keep a Changelog](https:
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
 
-## Version 0.8.4 - Unreleased
+## Version 0.8.5 - Unreleased
+
+### Added:
+* `kwimage.Affine.to_shapely` method for converting affine transforms to shapely-style 
+
+### Fixed
+* Fix issue in `Polygon.draw` where facecolor and edgecolor were not respected.
+* Added `to_shapely` for `kwimage.Affine`
+* Issue in `kwimage.load_image_shape` when using pathlib.Path
+
+
+## Version 0.8.4 - Released 2022-04-02
 
 ### Fixed
 * Bug in `Detections.draw_on(..., color='classes')` when a class is unknown